Match Unicode ranges before attempting binary search.
This changeset takes the ranges defined in the unicode table
and performs explicit checks before attempting the binary
search. This is needed because otherwise the binary search
will not find a value in a range since it is treated as a scalar.

Note that the generated table still contains the range identifiers,
but since its just a couple of them it doesn't make much sense
to explicitly exclude them.